Update backend testing + Simplify testing code + Upgrade backend to use node 20 + Update vitest to v1 and refactor accordingly + Update various packages
