+++ .github/workflows/backend-deploy-release.yml
@@ -8,12 +8,12 @@ on:
                 type: string
             node-version:
                 required: true
-                default: 18.x
+                default: 20.x
                 description: Node version to use
                 type: choice
                 options:
                     - 20.x
-                    - 18.x
+                    - 21.x
             upload-coverage:
                 required: false
                 default: true
