+++ .github/workflows/backend-test.yml
@@ -4,7 +4,7 @@ on:
     inputs:
       node-version:
         required: true
-        default: 18.x
+        default: 20.x
         type: string
       upload-coverage:
         required: false
