+++ .github/workflows/backend.yml
@@ -10,12 +10,12 @@ jobs:
   build-backend:
     uses: ./.github/workflows/backend-build.yml
     with:
-      node-version: 18.x
+      node-version: 20.x
       upload-artifact: true
   test-backend:
     uses: ./.github/workflows/backend-test.yml
     with:
-      node-version: 18.x
+      node-version: 20.x
   deploy-backend:
     if: github.event_name == 'push' && github.ref == 'refs/heads/develop' 
     needs: [build-backend, test-backend]
