+++ server/package.json
@@ -8,7 +8,7 @@
     "start": "node build/index.js",
     "startts": "npx ts-node src/index.ts",
     "dev": "npx nodemon src/index.ts",
-    "test": "vitest run --threads=false --reporter=verbose --coverage",
+    "test": "vitest run --pool=forks --poolOptions.forks.singleFork --coverage",
     "db": "docker-compose -f test_db/docker-compose-test-db.yml -p mikane_test_db up -d"
   },
   "dependencies": {
@@ -16,32 +16,35 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "helmet": "^7.0.0",
-    "nodemailer": "^6.9.5",
+    "helmet": "^7.1.0",
+    "nodemailer": "^6.9.7",
     "pg": "^8.11.3",
     "swagger-ui-express": "^5.0.0"
   },
   "devDependencies": {
-    "@types/cors": "^2.8.14",
-    "@types/express": "^4.17.18",
-    "@types/express-session": "^1.17.8",
-    "@types/node": "^18.18.3",
-    "@types/nodemailer": "^6.4.11",
-    "@types/pg": "^8.10.3",
-    "@types/supertest": "^2.0.14",
-    "@types/swagger-ui-express": "^4.1.4",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
-    "@vitest/coverage-v8": "^0.34.6",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/express-session": "^1.17.10",
+    "@types/node": "^20.10.4",
+    "@types/nodemailer": "^6.4.14",
+    "@types/pg": "^8.10.9",
+    "@types/supertest": "^2.0.16",
+    "@types/swagger-ui-express": "^4.1.6",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "@vitest/coverage-v8": "^1.0.4",
     "copyfiles": "^2.4.1",
-    "eslint": "^8.50.0",
-    "nodemailer-mock": "^2.0.1",
-    "nodemon": "^3.0.1",
+    "eslint": "^8.55.0",
+    "nodemailer-mock": "^2.0.2",
+    "nodemon": "^3.0.2",
     "rimraf": "^5.0.5",
     "supertest": "^6.3.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.2.2",
-    "vitest": "^0.34.6"
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
+    "vitest": "^1.0.4"
+  },
+  "engines": {
+    "node": "20.x.x"
   },
   "license": "GPL-3.0-or-later"
 }
